Scraping Tag-less Data and Making It Useful
Sometimes, scraping data can be really easy but this ease might translate to a lot of headache later due to absence of proper tags.
I was approached by a firm to scrape data from a particular website. They required Name, Address and Contact Info of the various contractors on that website. 
Here, I present an example in which I was easily able to scrape the data from that website. However, the sorting of the data was challenging due to the lack of proper tags in the html code. By removing some elements that were not necessary, I was able to find a pattern which was helpful in creating indexes to transform the data into something useful for business.